@@ -17,7 +17,7 @@ const Post = ({data}) => {
             <img src={Comment} alt="" />
             <img src={Share} alt="" />
         </div>
-        <span>{data.likes}Likes</span>
+        <span style={{color:"var(--gray)", fontSize:'12px'}}>{data.likes} Likes</span>
 
         <div className="detail">
             <span><b>{data.name}</b></span>
